@@ -1,4 +1,4 @@
-import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:wt_logging/wt_logging.dart';
 
@@ -15,14 +15,14 @@ abstract class LocalStorageStateNotifier<T> extends StateNotifier<T> {
     load();
   }
 
-  void load() async {
+  Future<void> load() async {
     final localStorage = await SharedPreferences.getInstance();
-    String? encodedValue = localStorage.getString(key);
+    final encodedValue = localStorage.getString(key);
     if (encodedValue == null) {
       log.d('Object has been loaded, but it was null. Setting site to "none".');
       state = none;
     } else {
-      T decodedValue = decode(encodedValue);
+      final decodedValue = decode(encodedValue);
       if (decodedValue == null) {
         log.d('Could not decode data from Key($key).');
       } else {
@@ -32,7 +32,7 @@ abstract class LocalStorageStateNotifier<T> extends StateNotifier<T> {
     }
   }
 
-  void replaceValue(T newValue) async {
+  Future<void> replaceValue(T newValue) async {
     log.d('Replacing data with Key($key): $newValue');
     state = newValue;
     final encodedValue = encode(state);
