@@ -1,9 +1,8 @@
 import 'package:flutter/widgets.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../base_settings_providers.dart';
-import 'settings_bool_componnet.dart';
-import 'settings_bool_notifier.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wt_settings/src/base_settings_providers.dart';
+import 'package:wt_settings/src/settings_bool/settings_bool_componnet.dart';
+import 'package:wt_settings/src/settings_bool/settings_bool_notifier.dart';
 
 class SettingsBoolProviders extends BaseSettingsProviders<SettingsBoolNotifier, bool> {
   final bool hideLabel;
