@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../base_settings_providers.dart';
-import 'settings_color_component.dart';
-import 'settings_color_notifier.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wt_settings/src/base_settings_providers.dart';
+import 'package:wt_settings/src/settings_color/settings_color_component.dart';
+import 'package:wt_settings/src/settings_color/settings_color_notifier.dart';
 
 class SettingsColorProviders extends BaseSettingsProviders<SettingsColorNotifier, MaterialColor> {
   final List<MaterialColor> values;
