@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wt_settings/src/settings_enum/settings_enum_notifier.dart';
+import 'package:wt_settings/src/settings_enum/settings_enum_providers.dart';
 
-import 'settings_enum_notifier.dart';
-import 'settings_enum_providers.dart';
-
-class SettingsEnumComponent<T extends Enum> extends HookConsumerWidget {
+class SettingsEnumComponent<T extends Enum> extends ConsumerWidget {
   final SettingsEnumProviders<T> providers;
 
   const SettingsEnumComponent({
@@ -30,15 +29,17 @@ class SettingsEnumComponent<T extends Enum> extends HookConsumerWidget {
             if (value != null) notifier.replaceValue(value);
           },
           items: providers.values
-              .map((e) => DropdownMenuItem(
-                    value: e,
-                    child: Text(_getTitle(e.toString())),
-                  ))
+              .map(
+                (e) => DropdownMenuItem(
+                  value: e,
+                  child: Text(_getTitle(e.toString())),
+                ),
+              )
               .toList(),
         ),
       ],
     );
   }
 
-  String _getTitle(String title) => title.contains('.') ? title.split(".")[1] : title;
+  String _getTitle(String title) => title.contains('.') ? title.split('.')[1] : title;
 }
