@@ -1,7 +1,6 @@
-import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:wt_logging/wt_logging.dart';
-
-import '../local_storage_state_notifier.dart';
+import 'package:wt_settings/src/local_storage_state_notifier.dart';
 
 class SettingsObjectNotifier<T> extends LocalStorageStateNotifier<T?> {
   static final log = logger(SettingsObjectNotifier);
@@ -26,15 +25,19 @@ class SettingsObjectNotifier<T> extends LocalStorageStateNotifier<T?> {
       if (_siteLoaded()) {
         if (_siteSelected()) {
           log.d(
-              'There is a current selected value, so checking if the current value is in the new list');
+            'There is a current selected value, so checking if the current value is in the new list',
+          );
         } else {
           if (_listLoaded(newList)) {
             if (_listContainsSelectedSite(newList)) {
               log.d('Selected value was in the new list');
             } else {
               log.d(
-                  'There is no current selected value, so first item of the list will be selected.');
+                'There is no current selected value, so first item of the list will be selected.',
+              );
               load();
+              // TODO: I suspect that the load here needs to be replaces with the following
+              //_setSiteToFirstSiteInList();
             }
           } else {
             log.d('There is no current selected value, but list has not been loaded yet.');
@@ -42,7 +45,8 @@ class SettingsObjectNotifier<T> extends LocalStorageStateNotifier<T?> {
         }
       } else {
         log.d(
-            'List has loaded, but the selected value has not loaded yet. Reloading selected value.');
+          'List has loaded, but the selected value has not loaded yet. Reloading selected value.',
+        );
         load();
       }
     });
@@ -66,6 +70,7 @@ class SettingsObjectNotifier<T> extends LocalStorageStateNotifier<T?> {
     return list != null && list.contains(state);
   }
 
+  // ignore: unused_element
   void _setSiteToFirstSiteInList(List<T>? list) {
     if (list != null && list.isNotEmpty) {
       super.replaceValue(list[0]);
