@@ -1,10 +1,9 @@
 import 'package:flutter/widgets.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../base_settings_providers.dart';
-import '../object_state_notifier.dart';
-import 'settings_object_componnet.dart';
-import 'settings_object_notifier.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wt_settings/src/base_settings_providers.dart';
+import 'package:wt_settings/src/object_state_notifier.dart';
+import 'package:wt_settings/src/settings_object/settings_object_componnet.dart';
+import 'package:wt_settings/src/settings_object/settings_object_notifier.dart';
 
 class SettingsObjectProviders<T> extends BaseSettingsProviders<SettingsObjectNotifier<T>, T?> {
   final StateNotifierProvider<StateNotifier<List<T>?>, List<T>?> listProvider;
